+++ pcs_test/tools/command_env/config_runner_pcmk.py
@@ -559,6 +559,7 @@ class PcmkShortcuts:
         stdout="",
         stderr="",
         returncode=0,
+        env=None,
     ):
         """
         Create a call for crm_resource --move
@@ -575,6 +576,7 @@ class PcmkShortcuts:
         string stdout -- crm_resource's stdout
         string stderr -- crm_resource's stderr
         int returncode -- crm_resource's returncode
+        dict env -- CommandRunner environment variables
         """
         # arguments are used via locals()
         # pylint: disable=unused-argument
@@ -669,6 +671,7 @@ class PcmkShortcuts:
         stdout="",
         stderr="",
         returncode=0,
+        env=None,
     ):
         cmd = ["crm_resource", action]
         if resource:
@@ -688,6 +691,7 @@ class PcmkShortcuts:
                 stdout=stdout,
                 stderr=stderr,
                 returncode=returncode,
+                env=env,
             ),
             before=before,
             instead=instead,